MOD structure - swap TV Shows with New Releases
- TV Shows have a slightly different response object structure compared
to the movie objects. To facilitate things in this demo project, the
TV Show page will therefore be swapped with the additional movie related
page. However, the TV Shows page and other related components will be
left in the project for future use.
- If the TV Shows page is needed at some point, a new search string
and a new details page will need to be created to meet the TV Show
criteria.